Cast to double for reliability

+++ backend/molgenis-emx2-datamodels/src/main/java/org/molgenis/emx2/datamodels/profiles/SchemaFromProfile.java
@@ -149,7 +149,7 @@ public class SchemaFromProfile {
             totalSizeEntry += nBytes;
             totalSizeArchive += nBytes;
 
-            double compressionRatio = totalSizeEntry / ze.getCompressedSize();
+            double compressionRatio = totalSizeEntry / (double) ze.getCompressedSize();
             if (compressionRatio > THRESHOLD_RATIO) {
               // ratio between compressed and uncompressed data is highly suspicious, looks like a
               // Zip
